moved status history chart into its own component
