+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-frontend/src/main/nifi/src/app/pages/summary/ui/processor-status-listing/processor-status-table/processor-status-table.component.ts
@@ -276,7 +276,7 @@ export class ProcessorStatusTable {
     }
 
     private compare(a: number | string, b: number | string, isAsc: boolean) {
-        return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
+        return (a < b ? -1 : a > b ? 1 : 0) * (isAsc ? 1 : -1);
     }
 
     private supportsMultiValuedSort(sort: Sort): boolean {
